reorganize lib files
